Data-Visualization-Dashboard/requirements.txt
pandas==2.0.0
numpy==1.24.0
sqlalchemy==2.0.0
python-dotenv==1.0.0
openpyxl==3.1.0
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.13.0
streamlit==1.22.0
chardet==5.1.0 